nitro: libkrun: fix macOS compilation failure
macOS was failing to compile with EFI=1 due to AWS nitro related code
compiling when it should not have been.
